+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -266,7 +266,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentKnowledgeBaseConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K6DAH8EYEHJDSWP7D87RGBMT"
+    id: Optional[str] = "01K6DHS9V2AD41SG7A0AQJJGCA"
     r"""The id of the resource"""