+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -670,5 +670,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-30T12:45:57.311Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-30T14:52:41.173Z")
     r"""The date and time the resource was last updated"""