+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -1291,5 +1291,5 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-30T12:45:57.311Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-30T14:52:41.173Z")
     r"""The date and time the resource was last updated"""